pyproject.toml
[tool.poetry]
name = "coke"
version = "0.1.0"
description = "Pytest-Alembic experiments (named coke because poetry is being weird)"
authors = ["Felix Fennell <felnne@bas.ac.uk>"]
license = "MIT"
readme = "README.md"
packages = [{include = "coke"}]

[tool.poetry.dependencies]
python = "^3.8"
pytest-alembic = "^0.9.1"
psycopg2 = "^2.9.5"
importlib-resources = "^5.10.2"  # backport
ulid-py = "^1.1.0"
geoalchemy2 = "^0.12.5"
shapely = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
